+++ Godeps/_workspace/src/github.com/jeromer/syslogparser/rfc3164/rfc3164.go
@@ -14,6 +14,7 @@ type Parser struct {
 	version  int
 	header   header
 	message  rfc3164message
+	location *time.Location
 }
 
 type header struct {
@@ -28,10 +29,15 @@ type rfc3164message struct {
 
 func NewParser(buff []byte) *Parser {
 	return &Parser{
-		buff:   buff,
-		cursor: 0,
-		l:      len(buff),
+		buff:     buff,
+		cursor:   0,
+		l:        len(buff),
+		location: time.UTC,
 	}
+}
+
+func (p *Parser) Location(location *time.Location) {
+	p.location = location
 }
 
 func (p *Parser) Parse() error {
@@ -137,7 +143,7 @@ func (p *Parser) parseTimestamp() (time.Time, error) {
 		}
 
 		sub = p.buff[p.cursor : tsFmtLen+p.cursor]
-		ts, err = time.Parse(tsFmt, string(sub))
+		ts, err = time.ParseInLocation(tsFmt, string(sub), p.location)
 		if err == nil {
 			found = true
 			break