+++ Godeps/_workspace/src/github.com/jeromer/syslogparser/syslogparser.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"strconv"
 	"strings"
+	"time"
 )
 
 const (
@@ -32,6 +33,7 @@ var (
 type LogParser interface {
 	Parse() error
 	Dump() LogParts
+	Location(*time.Location)
 }
 
 type ParserError struct {